// src/enum/stage.enum.ts
export enum StageState {
  PROSPECT = "Prospect",
  QUALIFIED = "Qualified",
  TECHNICAL_VALIDATION = "Technical Validation",
  BUSINESS_VALIDATION = "Business Validation",
  COMMITTED = "Committed",
  LAUNCHED = "Launched",
  CLOSED_LOST = "Closed Lost",
  CLOSED = "Closed",
  LOST = "Lost",
  WON = "Won",
  CREATED = "Created",
}

export enum GcpStage {
  DRAFT = "DRAFT",
  SUBMITTED = "SUBMITTED",
  ACCEPTED = "ACCEPTED",
  NOT_ACCEPTED = "NOT_ACCEPTED",
}